Move Android-specific code out of main_activity
https://phabricator.endlessm.com/T34155
